[249] Split Delete Button from Delete Dialog
